+++ modules/log/lab_test/farm_lab_test.post_update.php
@@ -176,3 +176,16 @@ function farm_lab_test_post_update_add_tissue_type(&$sandbox) {
   ]);
   $type->save();
 }
+
+/**
+ * Override the lab_test log timestamp label and description.
+ */
+function farm_lab_test_post_update_override_lab_test_timestamp_label_description(&$sandbox) {
+
+  // Override the timestamp field label and description on lab_test logs to
+  // make it clear that it should be used for the date the sample was taken.
+  $config = \Drupal::service('entity_field.manager')->getBaseFieldDefinitions('log')['timestamp']->getConfig('lab_test');
+  $config->set('label', 'Date sampled');
+  $config->set('description', 'The date when the sample was collected.');
+  $config->save();
+}